Attribute Option Entity removed
Removed AttributeOptions Entity simply because it is no entity in the
magento way. Rebuild most of the entity features into the
Import/Attribute/Option class to be able to use the known workflow. In
the future there has to be a massive overhaul of the class, with
outsourcing the entity fake methods to something like an Import/Abstract
to allow a more generalized way of importing non-entites (Cms/Rules come
to my mind).

The code is rather hacky, especially the Import.php additions, because
we need a real entity/non-entity importer detection in there to allow a
more generalized approach.