Set correct category position if not included in data
During initCategories, the position of the next child for every category is retrieved from the db and updated during the
import. If the position is not in the import data, the next_child_position is used and updated.

When an existing category is moved from one parent to another, it also gets a new position