SR-850 catch exception to ensure re-enable of config flag
SR-850 if the importer breaks at creating events or somewhere between, the store config value of AsyncIndex is disabled and therefore are not allowed to run by cronjob. This catch block ensures that it may run on error. If it breaks from php somewhere else or runtime is over for whatever reason, this fix wont stop that!